requirements.txt
numpy>=1.24,<2.0
pandas>=1.5,<2.1
matplotlib>=3.7,<3.9
scikit-learn>=1.2,<1.4
yfinance>=0.2.18,<0.3
streamlit>=1.28,<1.36

# Optional (code must gracefully fallback if unavailable)
statsmodels>=0.13,<0.15
tensorflow-cpu>=2.11,<2.14